summarization-service/requirements.txt
cohere
flask
python-dotenv